pyproject.toml
[project]
name = "umamusume-qq-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "python-dotenv>=1.2.2",
    "qq-botpy>=1.2.1",
]

[project.scripts]
umamusume-qq-bot = "umamusume_qq_bot.__main__:main"
